fix(web): add social sign-in/up buttons to Create Account screen
- Extract SocialButton to shared component (auth/SocialButton.tsx)
- Add social button grid to SignUp screen ("Or sign up with")
- Update SignIn to import from shared SocialButton
