+++ apps/web/src/components/auth/SignIn.tsx
@@ -6,6 +6,7 @@ import { GoogleIcon }    from '../icons/GoogleIcon';
 import { MicrosoftIcon } from '../icons/MicrosoftIcon';
 import { FacebookIcon }  from '../icons/FacebookIcon';
 import { AppleIcon }     from '../icons/AppleIcon';
+import { SocialButton }  from './SocialButton';
 
 interface SignInProps {
   onSuccess: () => void;
@@ -153,36 +154,5 @@ export function SignIn({ onSuccess, onSignUp, onForgotPassword, onConfirmRequire
         </button>
       </p>
     </AuthCard>
-  );
-}
-
-interface SocialButtonProps {
-  label: string;
-  bg: string;
-  textColor: string;
-  onClick?: () => void;
-  disabled?: boolean;
-  title?: string;
-  children: React.ReactNode;
-}
-
-function SocialButton({ label, bg, textColor, onClick, disabled, title, children }: SocialButtonProps) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      disabled={disabled}
-      title={title}
-      className="flex items-center justify-center gap-2 rounded-lg py-2 px-3 text-xs font-medium transition-opacity hover:opacity-90 disabled:opacity-50 disabled:cursor-not-allowed"
-      style={{
-        background:  bg,
-        color:       textColor,
-        fontFamily:  'var(--font-body)',
-        border:      '1px solid transparent',
-      }}
-    >
-      {children}
-      <span>{label}</span>
-    </button>
   );
 }
